openclaw/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openclaw"
version = "0.1.0"
description = "Multi-Platform Video Insight Skill System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=23.0",
    "rich>=13.0",
    "yt-dlp>=2024.0",
    "faster-whisper>=1.0",
    "semantic-text-splitter>=0.13",
    "chromadb>=0.4",
    "aiohttp>=3.9",
    "aiosqlite>=0.19",
    "asyncpg>=0.29",
    "pyyaml>=6.0",
    "cryptography>=41.0",
]

[project.optional-dependencies]
pdf = [
    "weasyprint>=60.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.0",
    "pytest-cov>=4.0",
]

[project.scripts]
openclaw = "openclaw.main:main"

[tool.hatch.build.targets.wheel]
packages = ["openclaw"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
